Moves persistent_subscription api to archive.h
The private types are now part of persistent_subscription.h

Now the placement of types and functions is consistent with
the rest of the system (e.g. replay merge).
