+++ aeron-archive/src/test/c/client/aeron_archive_persistent_subscription_context_test.cpp
@@ -20,8 +20,8 @@
 
 extern "C"
 {
+#include "client/aeron_archive.h"
 #include "client/aeron_archive_persistent_subscription.h"
-#include "client/aeron_archive_persistent_subscription_internal.h"
 }
 
 class AeronArchivePersistentSubscriptionContextTest : public testing::Test
