+++ aeron-archive/src/test/c/client/aeron_archive_persistent_subscription_test.cpp
@@ -34,7 +34,6 @@ extern "C"
 #include "concurrent/aeron_logbuffer_descriptor.h"
 #include "client/aeron_archive.h"
 #include "client/aeron_archive_persistent_subscription.h"
-#include "client/aeron_archive_persistent_subscription_internal.h"
 #include "uri/aeron_uri_string_builder.h"
 #include "util/aeron_env.h"
 #include "protocol/aeron_udp_protocol.h"
